vec: use uint32_t for array lengths
size_t should be used to represent sizes in bytes and is 64 bits. We
will never have vectors of that amount of elements. Use uint32_t for
array length and capacity.

+++ modules/infra/control/worker_test.c
@@ -118,10 +118,10 @@ mock_func(void *, __wrap_rte_zmalloc(char *, size_t, unsigned));
 #define assert_qmaps(qmaps, ...)                                                                   \
 	do {                                                                                       \
 		struct queue_map __expected[] = {__VA_ARGS__};                                     \
-		size_t __len = sizeof(__expected) / sizeof(struct queue_map);                      \
+		uint32_t __len = sizeof(__expected) / sizeof(struct queue_map);                    \
 		if (gr_vec_len(qmaps) != __len)                                                    \
-			fail_msg("%s len %zu expected %zu", #qmaps, gr_vec_len(qmaps), __len);     \
-		for (unsigned __i = 0; __i < __len; __i++) {                                       \
+			fail_msg("%s len %u expected %u", #qmaps, gr_vec_len(qmaps), __len);       \
+		for (uint32_t __i = 0; __i < __len; __i++) {                                       \
 			struct queue_map *exp = &__expected[__i], *act;                            \
 			bool found = false;                                                        \
 			gr_vec_foreach_ref (act, qmaps) {                                          \
